refactor: split subway station module from `Stop`
