+++ test/screens/v2/widget_instance/reconstructed_alert_property_test.exs
@@ -15,7 +15,7 @@ defmodule Screens.V2.WidgetInstance.ReconstructedAlertPropertyTest do
   alias ScreensConfig.V2.Header.CurrentStopId
   alias Screens.LocationContext
   alias Screens.RoutePatterns.RoutePattern
-  alias Screens.Stops.Stop
+  alias Screens.Stops.{Stop, Subway}
   alias Screens.Util
   alias Screens.V2.CandidateGenerator
   alias Screens.V2.WidgetInstance.ReconstructedAlert
@@ -1121,7 +1121,7 @@ defmodule Screens.V2.WidgetInstance.ReconstructedAlertPropertyTest do
       {:ok, now_datetime} = DateTime.from_unix(1_659_299_615)
 
       route_ids_at_stop =
-        Stop.get_all_routes_stop_sequence()
+        Subway.all_stop_sequences()
         |> Enum.filter(fn {_k, v} ->
           Enum.any?(v, fn sequence ->
             Enum.any?(sequence, fn stop_object ->
@@ -1143,7 +1143,7 @@ defmodule Screens.V2.WidgetInstance.ReconstructedAlertPropertyTest do
         end)
 
       tagged_station_sequences =
-        Map.new(route_ids_at_stop, fn id -> {id, [Stop.get_route_stop_sequence(id)]} end)
+        Map.new(route_ids_at_stop, fn id -> {id, [Subway.route_stop_sequence(id)]} end)
 
       station_sequences = RoutePattern.untag_stop_sequences(tagged_station_sequences)
